Add internal payment meta key to store number of child subscriptions
This is intended to be used only for past payments that were not created
correctly via admin and their payment items had the count of "1".
In the past the family subscriptions always had count of "1" and the real
count was only written in a payment note. This causes internal reporting
and processes to be confusing since the counts don't match and calculated daily
subscription prices are off.

remp/crm#2194