Fix not enought requests issue during family renewal
Originally we checked whether the amount of new subscriptions is higher
than the amount of old subscriptions (grouped by subscription type ID).

This was only partially correct because in order to renew
the subscription, we need to make sure that we have enough old
subscriptions to activate. This change iterates the "old" counts
and checks whether the new subscription contains enough of them.

remp/crm#2699